Preliminary support for Thumbnailer.lua
Even though it's preliminary, it seems to work quite well!
Thanks to deus0ww for doing all the heavy lifting on the way thumbnails
are handled and displayed. It's pretty much plug and play to support
other custom OSCs.

As for the next part: rewrite to reach parity with mpv's default OSC...
Not sure how long this is gonna take, since mordenx.lua seems to be
far behind mpv's own osc.lua.

Compare:
https://github.com/cyl0/mpv-osc-morden-x/commits/main/mordenx.lua
https://github.com/mpv-player/mpv/commits/master/player/lua/osc.lua