Rename mordenx.lua to modernx.lua
Since this fork is substantially different than the previous two forks,
it would be appropriate to name it somewhat differently too. How about
modernx? It's a one letter change, but it's enough to convey that this
fork strives to be divergent from its antecedent.

Jokes aside, yeah, I'm renaming it to the correct spelling. Not sure if
the original author didn't realise it's a typo or it's intentional, but
modern is the correct spelling.