Fix window controls when it is left-aligned
Since I increased the space between each window controls by 5px when it's
right-aligned, I've to compensate 10px if it's left aligned; so that the
close button doesn't get cut off. I'm not sure if there's a cleaner way
for this, since I do not want to change the close button position.