Fix crash if backspaced at the last position
Yes, this commit is just removing uncommented code, and here's why you
should: https://kentcdodds.com/blog/please-dont-commit-commented-out-code
:) But yeah, I made this fix before I learned about version control and
created this git repo. Uncommented code is bad and clutters the view.

As for why it was crashing, it's because current_time doesn't exist :D
I'm not sure why Lua is still checking it even though it's optional with
an "or", but removing it is all it takes to get rid of the error.