Remove superfluous cursor_position check for backspace()
Originally, the cursor_position checked is ~= 9. If the cursor is at the
9th position (which is the last), shift_cursor() won't kick in, aka the
cursor won't move to the left.

I don't like this behaviour, backspace should always shift the cursor to
the left even at the last position. My workaround is to increase the
check to ~= 10, which checks if the cursor_position is NOT 10th. There
are only 9 position, so the cursor will never reach the 10th position,
therefore the check is superfluous.