Make backspace() delete number first
A simple spell but quite unbreakable. This is finally the behaviour I
want! The main backspace() function, which is to "delete" (reset the
current number to 0) is now evaluated before the shift_cursor() function.
This is especially important if the current cursor_position is at the
last position. Now, the last position will be cleared before shifting to
the left.

Before: 00:05:42.069 -> 00:05:42.009
                                  ^ cursor position is here after
                                    backspacing at the last position
                                  v
 After: 00:05:42.069 -> 00:05:42.060