Fix timecode hitbox
Currently, the timecode isn't clickable at the seconds position (or the
hours position if it's on the right). This fix increases the width of
the timecode hitbox, allowing us to click the timecode up until the
millisecond position. Do note the hitbox don't dynamically increase or
decrease depending on whether milliseconds are shown, so there will be
an invisible clickable space beside the timecodes if milliseconds are
hidden.