Use `current-window-scale` to reset to native resolution
Not sure when this was fixed, but `current-window-scale` now works.

From https://mpv.io/manual/master/#command-interface-window-scale,

"Setting a value identical to its previous value will not
 resize the window."

That's why setting `window-scale` to 1 previously didn't work, because
it already started with the default of 1, so we have to use the hacky
method of setting it to 1.001. However, this hack only works once, since
we already set the `window-scale` to 1.001, if we set it again, it will
not resize the window.

The solution here is to use `current-window-scale`, which previously
didn't work. Now it does work pretty reliably.