Fix crash when toggling Thumbnailer's OSC before the OSC is shown
This only happens on a freshly launched mpv before the OSC is shown.
If you open a media file and immediately toggle Thumbnailer's OSC
("Thumbnailer-toggle-osc" in input.conf), there's a chance the script
will crash. If the OSC is shown at least once, then the crash will not
happen. It might be because mpv haven't determined the OSC "state" when
it's just launched.

Temporary workaround for deus0ww/mpv-conf#17