Remove `toggle_osc.lua`
`toggle_osc.lua` was originally a workaround for the inability to toggle
osc-visibility with DEL. Now that's fixed in 967413b, I no longer need
this script anymore. I can quickly show the OSC with DEL, hide the OSC
with shift+DEL and peek the "OSC" with 'o', so the functionalities are
all there.