Improve handling of auto-delete with multiple instances
This commit improves the handling of Thumbnailer.lua's auto-delete
thumbnail cache feature when there are multiple instances of the same
video playing. Basically, a lock will be created in the thumbnail folder
when a video is playing. When the video finishes playing, the lock will
be released and the thumbnails folder is deleted as usual.

Assuming both videos are the same file and auto-delete is enabled, this
fix prevents the 2nd video from losing its thumbnails when the 1st video
is closed.

Relevant issues: deus0ww/mpv-conf#15 hooke007/MPV_lazy#48
Original commit: hooke007/MPV_lazy@273969f
Credit goes to the original author, of course.