Fix race condition when detecting lock before folder is created
In reality, this race condition should rarely happen. The race condition
occurs if Thumbnailer.lua tries to delete the thumbnail folder before it
is created. This is only possible if you open and quit mpv quickly.

The only reason I'm obliged to fix this is because I despise seeing the
stack trace gets printed every time I open a file by accident and have
to quit it immediately.