Improve modernx OSC for vertical videos
Partly mentioned in #2. Currently, the OSC icons will clash each other
when the video is narrow and vertical. This commit simplifies the OSC
and hide extra icons when there's not much space available.

When the video played is <480p in width and <1 in aspect ratio:
- The "previous/next playlist", "previous/next chapter", "toggle info"
  buttons and cache indicator are hidden
- The volume icon is moved to the right and replaces "toggle info"
- The audio and subtitle icon remains on the left
- The seek backward/forward icons will have a slight padding decrease

Of course, there's a limit on how narrow the video can be before the
icons start clashing again. This should at least work for vertical
videos with an aspect ratio of 9:16.