Enable `--video-sync=display-resample` for low-FPS videos
`display-resample` tries to output video frames completely synchronously
to the display. In other words, the OSC will render in 60FPS (or
whatever your display's refresh rate is) instead of being chained to the
video framerate. This really helps to alleviate the sluggishness of the
OSC in low-framerate videos.

That said, `display-resample` isn't as robust as the default `audio`.
The way it stutters (with audio) is very distracting. Without audio, it
tries to render every frame (like a compositing software) instead of
dropping frames to get realtime playback. It also consumes a negligible
extra GPU power. `display-resample-vdrop` didn't make much of a
difference.

Additionally, a quick glance at mpv's issue tracker reveals a slew of
issues related to display-resample. Therefore, I've created a profile
condition that only applies to <30FPS videos, where it actually matters.

YMMV, more testing is needed.