Add feature to show OSC on `file-loaded` and `seek`
Showing the OSC on `file-loaded` helps to indicate that the current file
has changed, e.g. when the next file in a playlist starts playing.
Whereas showing OSC on `seek` is useful for users who prefer seeing the
full OSC (instead of the minimal osd-bar) when seeking.

Of course, these features can be toggled in osc.conf.

Resolves #4