Remove SmartCopyPaste_II
With the release of SmartCopyPaste 3.1, there's finally an option to
copy the timestamp with milliseconds precision, along with other QOL
features like multi-paste.

The only hurdle to a smooth upgrade is that the author took a rather
weird development approach: instead of adding features incrementally,
they opted for a release-esque workflow, which makes updating harder.
Additionally, for some reasons there are two separate scripts for
SmartCopyPaste, and the author decided to only update one of them.
¯\_(ツ)_/¯

Anyways, cut the rant; To reduce conflicts, I'm going to remove the
current SmartCopyPaste_II and start over.