Round off `osc_param.display_aspect`
This fixes two special cases:
- Icons not appearing in specific aspect ratios.
  e.g. The playlist buttons don't appear at an aspect ratio of 1.1.
- OSC elements frantically shows/hides when resizing the window bigger,
  most apparent for videos with an aspect ratio of 0.6.

These glitches only happen when the aspect ratio is exactly the same as
the value being compared to, i.e. 1.3, 1.1, 0.9 and 0.6.

That's because `osc_param.display_aspect` is only a near approximation
of the true aspect ratio. For example, in windowed mode, a video with an
aspect ratio of 0.6 will yield "0.599771...", thus it will be considered
smaller than 0.6, resulting in some icons not showing as they should.

The 2nd issue is that when you resize the window *bigger* (not smaller),
`osc_param.display_aspect` fluctuates within ±0.001 of the actual aspect
ratio, enough to cause some OSC elements popping up for a split second.

Therefore, the solution here is to round `osc_param.display_aspect` to 1
decimal. Attribution for the code: http://lua-users.org/wiki/SimpleRound