Make window controls respect Fitts's law
Increase the geometry of the window controls to match the height of
wcBar, so that the close button is clickable when you move the cursor to
the top right corner in fullscreen.

I don't plan on doing the same with the volume button and fullscreen
button since they are too far away from their respective corners.

Originally inspired from mpv-player/mpv#9791