Observe and save runtime changes to `sub-pos`
As a last safeguard against interfering with the user's sub-pos,
the script will now remember any runtime changes to `sub-pos`,
whether via keybinds or other means.

Note that the script will only keep track of changes when the OSC is
not visible. That's because when the OSC is visible, update_subpos()
will take effect. We DON'T want to accidentally remember the sub-pos
_after_ it's been moved up.