Expose `osc-margins` to move the console
Originally from mpv-player/mpv@4787eb5. I tried to keep the margins
implementation as minimal as possible, since I don't plan to add
"boxvideo" as a feature.

I also disagree with the assumption that "it would look weird to move
the console depending on the mouse movement", so I have removed one of
the checks that prevented so.

Other included upstream commits:
- mpv-player/mpv@0d38459
- mpv-player/mpv@530b203