Only use `video-sync=display-resample` on videos below 4K
On one of my 4K samples, display-resample actually made the OSC less
smooth, i.e. the OSC wouldn't disappear or show up at all when the video
is playing. According to the stats page, <something> (the rendering
thread?) is starving the refresh driver, causing the video-sync to be
delayed accumulatively, to the point that the OSC doesn't get the chance
to update itself. Since it's rare for a 4K video to be less than 24FPS,
just disable the auto-profile condition for >4K.