Refactor hover effect code
- Revert commit 1da953337a4c5005043017ba14c894ab9e6ad0a2

That commit has always been a stopgap measure because I can't find the
reason why the hover effect removes the elementDown style for vidTitle.

According to https://aegi.vmoe.info/docs/3.1/ASS_Tag, the "\blur" tag
will only blur the text itself (emulating the glow effect) if there's no
border, hence "\bord0" is required. That's possible for the OSC buttons
and timecodes, but not feasible for vidTitle as it has to stand out.
The solution is to set the secondary fill colour using the "\2c" tag.

This does come with a minor unintended side effect:
- The glow/hover effect gets affected by the elementDown style

Other changes grouped in this commit:
- Add hover effect to vidTitle
- Fix visible clipping at the beginning and below the title on hover
- Fix hover effect on window controls button (they now glow)
- Use Zren/mpv-osc-tethys@83df9d4 approach to disable hover effect on
  unclickable elements
- Reduce vidTitle height, while still giving some leeway for fonts with
  taller characters. This is mainly to avoid clicking vidTitle if the
  cursor is far above it