Tweak ffmpeg threads
For my CPU model, my testing shows that 4 workers 2 threads generates
thumbnails the quickest while using the least RAM. See the table below:

| Samples  |    Workers/Threads    |
| -------- | 4/8 | 4/4 | 4/2 | 8/2 |
| H.264    | 13s | 13s | 10s | 11s | - 4/2
| HEVC     | 37s | 33s | 29s | 27s | - 8/2
| VP9      | 51s | 47s | 47s | 45s | - 8/2
| AV1      | 77s | 73s | 71s | 73s | - 4/2

Note: Each sample have varied durations and resolutions, so no
conclusions regarding which video formats are easier to decode can be
drawn. They merely serves as labels.

From the table, 4/2 and 8/2 are tied as being the fastest. Picking 4/2
would be a no-brainer as it's also ligher on RAM. Keep in mind that the
default of 3 workers 8 threads performed the worst, so it's important to
tweak the workers/thread options to take full advantage of Thumbnailer.