Tweak mpv workers & threads
Unlike ffmpeg, mpv with the default setting of 3 workers and 8 threads
does not always 100% utilise the CPU, giving the impression that it is
slower. With further tweaking, it's possible for mpv to completely use
the CPU, so that it's fairer to compare to ffmpeg.

| Samples  |    Workers/Threads    |
| -------- | 4/8 | 8/4 | 8/2 | 6/2 |
| H.264    | 18s | 10s | 10s | 12s | - 8/2
| HEVC     | 45s | 37s | 29s | 31s | - 8/2
| VP9      | 69s | 60s | 57s | 60s | - 8/2
| AV1      | 24s | 17s | 16s | 16s | - 6/2 (???)

There we go, the thumbnail generation speed has increased considerably,
much more so than after tweaking ffmpeg. Due to the tendency of mpv
being less multi-threaded, it's advisable to set the max_workers to
the number of cores in your CPU.

Interestingly, mpv significantly outperforms ffmpeg in generating
thumbnails for AV1 videos. The reason is unknown, much like how ffmpeg
decodes the VP9 sample quicker than mpv can.