Implement logarithmic volume control
Upstream commit: maoiscat/mpv-osc-modern@42174e2

The volume scale is actually exponential, but it was named that way
because mpv does not use db as the audio unit. I try to be consistent
with other music players (e.g. foobar2k) calls that option.

As quoted from https://www.dr-lex.be/info-stuff/volumecontrols.html,

    We are much more sensitive to small variations in amplitude for
    silent sounds than for loud sounds

which is why I used an exponential scale, as opposed to maoiscat's
logarithmic scale.