crop: Improve click-and-drag crop selection
Second take of d5306ca because I noticed Sneakpeakcss didn't fully
implement click-and-drag crop selection exactly as in the original PR.
Previously, you had to click the first point before you can hold and
drag to crop.

This commit integrates po5's complex keybinds changes, allowing you to
start, select and submit the crop in a single hold-and-drag operation!

One of the roadblock to patching po5's PR directly was the condition
`or (crop_first_corner.x == corner_video.x and crop_first_corner.y ==
corner_video.y)`. It was causing errors due to a missing `corner_video`
variable, which was removed in occivink/mpv-scripts@af360f3. Thus, I've
partially restored the variable to fix this.

Note: Unlike Sneakpeakcss's version, click-and-drag and point-and-click
are mutually exclusive. You can't click to start the crop, then
hold-and-drag to select the area. It was unintuitive anyways.

Ref: occivink/mpv-scripts#62