Update config
- Don't reset "mute" on next file

Turns out that even reloading will mute the video, which is more
aggresive than intended. Though, this returns the unwanted behaviour:
If the current video is not muted and is playing, mpv will auto-play
the next video with sound. As a workaround, I can just pause or mute the
current video before moving on to the next video.

- Update playlistmanager style ASS tags

This is finally the appearance I'm after! It's not really possible to
blur only the shadow and not the border, without resorting to layering
techniques. Thus, I've toned down the border and blur, then add a drop
shadow effect to fake the blur.