Add `play-dir` keybind
This commit reverts 6fba410, but also adds a fun little keybind for
`play-dir`. It's useful for edge cases where I want to seek backward
quickly but less than 1s. I removed the previous keybind because I've
found "seek -0.05 exact" to be unrealiable. Anyways, I've identify the
cause for slow back-step, which will be explained in the next commit.