Update console font size and outline width
It is a somewhat radical change, though I have largely become accustomed
to it. The size may be a little large in windowed mode, but otherwise
looks fine in full-screen mode. With these two options, the console text
should be significantly more readable and sharper, even against the
noisiest background.