Update mpv internal profiles
With the changes in default settings courtesy of mpv-player/mpv@703f158,
mpv now strives for a balance between quality and speed, rather than the
highest performance by default. Previously, I have to opt-in the
"high-quality" profile, but now it's an opt-out. However, it's not as
simple as swapping the profiles keybind in my config. The `mid` profile
was created to incorporate `profile-restore=copy` to make it toggleable.

[background-throttling] has also been modified to take advantage of
profile=fast, with a few additions. If you're wondering why
`correct-downscaling=yes` and `dscale=hermite` is here, I also use this
profile for PiP purpose. Without these two options, the video looks
atrocious when downscaled.