Add profile-restore to `delete-watch-later-config` autoprofile
This ensures that `save-position-on-quit` is disabled ONLY on EOF, and
re-enabled as soon as the user seeks to other parts of the file.
Previously, if a user seeks to EOF, then seek to the middle of the file,
the position won't be saved. That's because once the player reaches EOF,
the autoprofile is triggered, `save-position-on-quit` will then be
disabled for the rest of the session.