conf: Fix incompatibility between skipsilence and dynaudnorm
When dynaudnorm is placed after skipsilence's lavfi graph in the filter
chain, skipsilence breaks and speeds up during non-silent sections.
Initially, I thought it was an incompatibility between dynaudnorm and
scaletempo2 because the console errors with:

  [lavfi] speed changing filters before libavfilter are not supported
          and can cause desyncs

However, this turned out to be a separate issue. The root problem lies
with skipsilence itself, which always prepends its filter graph.
Consequently, dynaudnorm is always applied after silencedetect,
regardless of when it's applied.

To circumvent this issue, I've created an auto-profile that re-prepends
dynaudnorm when skipsilence is enabled. This also resolves the afore-
mentioned error by ensuring dynaudnorm is placed before scaletempo2.
Note that the error message may not be entirely avoided due to a
slight delay in applying the auto-profile.