conf: Simplify profile-cond for `reorder-dynaudnorm`
Here, I'm going to document on handling properties that are natively
returned as tables in profile-cond. Looking at the docs for `af` [1],
I was misguided that `mp.get_property_native()` was required to access
the key-values pair in the `af` property. This lead me to iterate over
the returned table to check for a specific filter. [2] Turns out it's
possible to query the `af` property as a string with `mp.get_property()`
which drastically simplifies the logic. [3][4]

For documentation sake, this is the previous profile-cond expanded:
    get("user-data/skipsilence/enabled") and (function()
        local af = mp.get_property_native("af")
        for _, filter in pairs(af) do
            if filter["name"] == "dynaudnorm" then
                return true
            end
        end
    end)()

[1]: https://mpv.io/manual/master/#command-interface-af
[2]: https://mpv.io/manual/master/#lua-scripting-mp-get-property-native(name-[,def])
[3]: https://mpv.io/manual/master/#lua-scripting-mp-get-property(name-[,def])
[4]: https://github.com/mpv-player/mpv/issues/12127