seek-to: Improve visual for selected digit
By default, seek-to uses a bold tag to indicate selected digit, which is
not helpful if the user already uses `osd-bold=yes`. This commit
repurposes the redraw timer to create an underline indicator for the
selected digit. The selected indicator can be customised through
`selection_color` and `selection_border_color`.

Credit: Sneakpeakcss/mpv-config-seek-to@9c5db4f
        Sneakpeakcss/mpv-config-seek-to@18562f8