seek-to: Fix timestamp shifting when using default font
The timestamp may visually shift when moving the selection, if the
timestamp contains "thin" number such as '1'. This commit fixes it by
adding a very thin space (aka "hair space") and forcing its scale to 0.

While the original commit forces bold font for the whole seek-to prompt,
I find it unnescessary as the supposed "empty space" caused by the hair
space is imperceptible.

As noted by the author, this was tested only on mpv's default font on
Windows, specifically Arial. For some reason, only some fonts (like
Arial) have this issue, even with tabular figures; other fonts (like
Manrope TabNum) do not. But well, if it works it works.

Credit: Sneakpeakcss/mpv-config-seek-to@9c5db4f