seek-to: Remember timestamp if exited before confirming
This reverts edcee30.

That commit introduced two unwanted behaviours:
- If seek-to is closed before the timestamp is submitted, it resets to 0
- The history_position is reset on close

I didn't like those behaviours, but that commit was needed to prevent it
from crashing after pasting timestamp. Now, I'm here to investigate and
restore the previous behaviour.

Previously, the deduplication code compared the timestamp of the current
selected entry to the last history entry. This didn't pose a problem as
the history_position is reset to last on close. I've corrected it to
compare the pasted entry to the last entry, restoring the 2nd behaviour.

After the script identifies duplication, a new entry will be initialized
as an empty table. This empty table is nil, causing the script to crash.
To restore the 1st behaviour, I've moved the code that populates the
empty table from set_inactive() to paste_timestamp().