conf: Replace SmartCopyPaste with copyStuff
Now that seek-to accepts both seconds and sexagesimal format as
timestamps, I no longer have to rely on SmartCopyPaste to handle it.
In fact, I never needed to because mpv's `seek` command would parse it
anyways, but the recent enhancement to seek-to had sealed the deal.
With that, I can fully use HH:MM:SS in my workflow for good.

copyStuff is a simple and straightforward script that fulfills my needs,
without the fluff from SmartCopyPaste. As for the paste function, well,
I can simulate it with mpv's built-in console command.