seek-to: Check if timestamp exceeds video length before seeking
The original intention of this feature is to prevent mpv from advancing
to the next file if the timestamp exceeds the video length. Honestly,
it's an overkill workaround for a problem caused by mpv's default
setting. The simpler fix is to set `keep-open=always` or pause the video
before seeking. But since it constitutes as an unexpected behaviour,
let's backport the changes.

Credits: Sneakpeakcss/mpv-config-seek-to@ad23e06