crop: Refactor remove_crop() out of toggle_crop()
Another approach to occivink/mpv-scripts#77
The code for removing the last crop filter is already in toggle_crop(),
I just have to break it down into smaller functions.

Firstly, remove_delogo() is invoked to remove the last delogo filter.
If a delogo filter exists, the function will remove it and return true;
remove_hard() won't be called.
Conversely, if no delogo filter is found, remove_hard() is invoked to
determine the state of video-crop.
Finally, if no delogo filter is present or video-crop is disabled,
toggle_crop() will call start_crop().

The order of crop removal will be made configurable in the next few
commits.