skipsilence: Use `alt_normal_speed` to fix clicks/gaps
Quote from the Recommendations[1]:

    mpv automatically inserts scaletempo2 filter when changing speed and
    removes it when returning to 1x playback speed. This often causes
    clicks in audio playback.
    ...
    In mpv 0.36 and below, the scaletempo2 filter itself additionally
    caused similar artifacts when switching to and from 1x speed.
    This has been fixed in mpv 0.37. The `alt_normal_speed` option of
    this script is a workaround for older versions.

To clarify, these are two separated issues with different factors. The
former is caused by mpv and it's still reproducible, while the latter is
caused by scaletempo2 and has been resolved. The README suggests
distinct solutions for these issues.

However, both problems occur when changing the speed to-and-fro 1x,
sharing the same pattern. Thus, the scenario addressed by
`alt_normal_speed` should theoretically apply to the former bug. With
this option, the scaletempo2 workaround is not necessary. Surprisingly,
1.001 speed seems sufficient to prevent clicks.

By eliminating the need for scaletempo2, I can indirectly resolve a
long-standing annoyance:
- I've a keybind to toggle skipsilence and scaletempo2 simultaneously
- After saving and reloading, skipsilence is disabled while scaletempo2
  remains active, causing the keybind to desync
- Although mpv saves the af chain on exit, the skipsilence script is
  internally disabled on each launch
- As a workaround, I have to either manually `af remove scaletempo2` to
  resync the keybind or disable skipsilence before quitting

[1]: https://github.com/ferreum/mpv-skipsilence?tab=readme-ov-file#fix-clicking-soundsgaps-when-switching-to-and-from-1x-speed