conf: Add ArtCNN_C4F8_DS.glsl
It's exciting to see a new upscaling shader competing with FSRCNNX!
I chose the C4F8 model because it's the only one that runs reasonably
on my system. However, I've only opted the DS version because the base
version don't offer a significant advantage over FSRCNNX, contrary to
the blog post. [1] Perhaps it's more "correct" or "faithful" to the
source, or more optimised for anime, but I find it a tad blurrier than
FSRCNNX_8-0-4-1 for my use cases.

Regardless, the DS version works as advertised, improving the experience
of watching poorly compressed sources by denoising and sharpening. While
FSRCNNX has an artifact-removal version too [2], it's only available in
16-0-4-1, which is unsuitable for real-time use. Thankfully, ArtCNN-DS
can now serve as a lighter alternative to FSRCNNX-Distort.

[1]: https://artoriuz.github.io/blog/mpv_upscaling.html
[2]: https://github.com/HelpSeeker/FSRCNN-TensorFlow/releases