crop: Disable window dragging when cropping
This lets users hold and drag to select a crop area. This is implemented
by temporarily setting `window-dragging` property to no when cropping.
Releasing the mouse confirms the crop. The behaviour for point-and-click
to crop is unchanged. This option can be disabled with a `script-opt`.

Recent changes to mpv caused issues detecting mouse input when selecting
an area. Disabling `window-dragging` resolves this problem.

Ref: https://github.com/Sneakpeakcss/mpv-scripts/commit/136d435