conf: Increase `demuxer-max-back-bytes` to 200MiB
This increases the total cache/buffer to 500MiB, which is useful for
larger network streams such as livestreams. I realised more often than
not, the backward buffer is just as important as the forward buffer if
you want to rewind some parts, so there's no reason to be stingy with
the cache size.