memo: Remove script
As hinted in b9ad214, the main function of this script - saving watch
history and displaying it in a menu - has been implemented in core mpv
through `select.lua`. The only unique features left are its integration
with uosc and directories history, both which are not needed.
That's one less script to maintain.