seek-to: Add different seek modes
"-" key seeks backward relative to current position
"+" key seeks forward relative to current position
"=" key seeks from the end of the video
The appropriate symbol will be shown when toggled.
Also adds their respective numpad keys.

The symbol for "seek from end" is "<", which is inconsistent with its
"=" keybind. The keybinds are chosen because "- + =" are grouped on the
keyboard. "<" should be easily understood in this context.

Credits: https://github.com/Sneakpeakcss/mpv-config-seek-to/commit/e0f3e18
         https://github.com/Sneakpeakcss/mpv-config-seek-to/commit/eb2713b